SOL-04 Cambiado el nombre del controlador
