+++ app/assets/stylesheets/service_requests.css.scss
@@ -1,3 +1,3 @@
-// Place all the styles related to the service controller here.
+// Place all the styles related to the service_requests controller here.
 // They will automatically be included in application.css.
 // You can use Sass (SCSS) here: http://sass-lang.com/
