--- app/helpers/service_helper.rb
@@ -1,2 +1,0 @@
-module ServiceHelper
-end
