+++ app/helpers/service_requests_helper.rb
@@ -1,0 +1,2 @@
+module ServiceRequestsHelper
+end
