--- test/controllers/service_controller_test.rb
@@ -1,9 +1,0 @@
-require 'test_helper'
-
-class ServiceControllerTest < ActionController::TestCase
-  test "should get requests" do
-    get :requests
-    assert_response :success
-  end
-
-end
