+++ test/controllers/service_requests_controller_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class ServiceRequestsControllerTest < ActionController::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
