--- test/helpers/service_helper_test.rb
@@ -1,4 +1,0 @@
-require 'test_helper'
-
-class ServiceHelperTest < ActionView::TestCase
-end
