+++ test/helpers/service_requests_helper_test.rb
@@ -1,0 +1,4 @@
+require 'test_helper'
+
+class ServiceRequestsHelperTest < ActionView::TestCase
+end
